move white furnace to the 3d testbed
